pyproject.toml
[project]
name = "paperweave"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.1.3",
    "black==23.10.0",
    "certifi==2024.8.30",
    "chainlit>=2.2",
    "jupyter>=1.1.1",
    "langchain>=0.3.17",
    "langchain-ollama>=0.2.1",
    "langchain-openai>=0.2.14",
    "langgraph>=0.2.35",
    "langgraph-cli>=0.1.54",
    "langsmith>=0.1.147",
    "networkx>=3.4.2",
    "openai>=1.61.1",
    "pydantic>=2.7.4",
    "pydantic-core>=2.18.4",
    "python-dotenv>=1.0.1",
    "scikit-image>=0.25.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "transformers>=4.46.3",
    "typer==0.13",
    "typing-extensions==4.12.2",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["paperweave"]

[tool.uv.sources]
paperweave = { workspace = true }

[tool.uv]
package = true
